fix flickering problem with the link data UI.
We have a flickering problem on zoom and when adding data due to the removal and creation of new paths.
One main reason is that the path drawingis made inside requestAnimationFrame which sometime call a loop of digest.

In order to keep the solution maintain, I extracted the datalinks object outside of the links and draw it a seprate part next to the lines.
which result in smooth drawing without deleting and recreating the objects.